+++ Tests/PackageLoadingTests/ManifestLoaderCacheTests.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift open source project
 //
-// Copyright (c) 2020-2021 Apple Inc. and the Swift project authors
+// Copyright (c) 2020-2023 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See http://swift.org/LICENSE.txt for license information
@@ -16,7 +16,323 @@ import PackageModel
 import SPMTestSupport
 import XCTest
 
+import class TSCBasic.InMemoryFileSystem
+import func TSCTestSupport.withCustomEnv
+
 class ManifestLoaderCacheTests: XCTestCase {
+    func testDBCaching() throws {
+        try testWithTemporaryDirectory { path in
+            let fileSystem = localFileSystem
+            let observability = ObservabilitySystem.makeForTesting()
+
+            let manifestPath = path.appending(components: "pkg", "Package.swift")
+            try fileSystem.createDirectory(manifestPath.parentDirectory, recursive: true)
+            try fileSystem.writeFileContents(
+                manifestPath,
+                string: """
+                    import PackageDescription
+                    let package = Package(
+                        name: "Trivial",
+                        targets: [
+                            .target(
+                                name: "foo",
+                                dependencies: []),
+                        ]
+                    )
+                    """
+            )
+
+            let delegate = ManifestTestDelegate()
+
+            let manifestLoader = ManifestLoader(
+                toolchain: try UserToolchain.default,
+                useInMemoryCache: false,
+                cacheDir: path,
+                delegate: delegate
+            )
+
+            func check(loader: ManifestLoader, expectCached: Bool) throws {
+                delegate.clear()
+                delegate.prepare(expectParsing: !expectCached)
+
+                let manifest = try XCTUnwrap(loader.load(
+                    manifestPath: manifestPath,
+                    packageKind: .root(manifestPath.parentDirectory),
+                    toolsVersion: .current,
+                    fileSystem: fileSystem,
+                    observabilityScope: observability.topScope
+                ))
+
+                XCTAssertNoDiagnostics(observability.diagnostics)
+                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1), [manifestPath])
+                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, expectCached ? 0 : 1)
+                XCTAssertEqual(manifest.displayName, "Trivial")
+                XCTAssertEqual(manifest.targets[0].name, "foo")
+            }
+
+            try check(loader: manifestLoader, expectCached: false)
+            for _ in 0..<2 {
+                try check(loader: manifestLoader, expectCached: true)
+            }
+
+            try fileSystem.writeFileContents(
+                manifestPath,
+                string: """
+                    import PackageDescription
+
+                    let package = Package(
+                        name: "Trivial",
+                        targets: [
+                            .target(
+                                name: "foo",
+                                dependencies: [  ]),
+                        ]
+                    )
+
+                    """
+            )
+
+            try check(loader: manifestLoader, expectCached: false)
+            for _ in 0..<2 {
+                try check(loader: manifestLoader, expectCached: true)
+            }
+
+            let noCacheLoader = ManifestLoader(
+                toolchain: try UserToolchain.default,
+                useInMemoryCache: false,
+                cacheDir: .none,
+                delegate: delegate
+            )
+            for _ in 0..<2 {
+                try check(loader: noCacheLoader, expectCached: false)
+            }
+
+            // Resetting the cache should allow us to remove the cache
+            // directory without triggering assertions in sqlite.
+            manifestLoader.purgeCache(observabilityScope: observability.topScope)
+            XCTAssertNoDiagnostics(observability.diagnostics)
+            try fileSystem.removeFileTree(path)
+        }
+    }
+
+    func testInMemoryCaching() throws {
+        let fileSystem = InMemoryFileSystem()
+        let observability = ObservabilitySystem.makeForTesting()
+
+        let manifestPath = AbsolutePath.root.appending(components: "pkg", "Package.swift")
+        try fileSystem.createDirectory(manifestPath.parentDirectory, recursive: true)
+        try fileSystem.writeFileContents(
+            manifestPath,
+            string: """
+                import PackageDescription
+                let package = Package(
+                    name: "Trivial",
+                    targets: [
+                        .target(
+                            name: "foo",
+                            dependencies: []),
+                    ]
+                )
+                """
+        )
+
+        let delegate = ManifestTestDelegate()
+
+        let manifestLoader = ManifestLoader(
+            toolchain: try UserToolchain.default,
+            useInMemoryCache: true,
+            cacheDir: .none,
+            delegate: delegate
+        )
+
+        func check(loader: ManifestLoader, expectCached: Bool) throws {
+            delegate.clear()
+            delegate.prepare(expectParsing: !expectCached)
+
+            let manifest = try XCTUnwrap(loader.load(
+                manifestPath: manifestPath,
+                packageKind: .root(manifestPath.parentDirectory),
+                toolsVersion: .current,
+                fileSystem: fileSystem,
+                observabilityScope: observability.topScope
+            ))
+
+            XCTAssertNoDiagnostics(observability.diagnostics)
+            XCTAssertEqual(try delegate.loaded(timeout: .now() + 1), [manifestPath])
+            XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, expectCached ? 0 : 1)
+            XCTAssertEqual(manifest.displayName, "Trivial")
+            XCTAssertEqual(manifest.targets[0].name, "foo")
+        }
+
+        try check(loader: manifestLoader, expectCached: false)
+        for _ in 0..<2 {
+            try check(loader: manifestLoader, expectCached: true)
+        }
+
+        try fileSystem.writeFileContents(
+            manifestPath,
+            string: """
+                import PackageDescription
+
+                let package = Package(
+                    name: "Trivial",
+                    targets: [
+                        .target(
+                            name: "foo",
+                            dependencies: [  ]),
+                    ]
+                )
+
+                """
+        )
+
+        try check(loader: manifestLoader, expectCached: false)
+        for _ in 0..<2 {
+            try check(loader: manifestLoader, expectCached: true)
+        }
+
+        let noCacheLoader = ManifestLoader(
+            toolchain: try UserToolchain.default,
+            useInMemoryCache: false,
+            cacheDir: .none,
+            delegate: delegate
+        )
+        for _ in 0..<2 {
+            try check(loader: noCacheLoader, expectCached: false)
+        }
+
+        manifestLoader.purgeCache(observabilityScope: observability.topScope)
+        XCTAssertNoDiagnostics(observability.diagnostics)
+    }
+
+    func testContentBasedCaching() throws {
+        try testWithTemporaryDirectory { path in
+            let manifest = """
+                import PackageDescription
+                let package = Package(
+                    name: "Trivial",
+                    targets: [
+                        .target(name: "foo"),
+                    ]
+                )
+            """
+
+            let delegate = ManifestTestDelegate()
+
+            let manifestLoader = ManifestLoader(
+                toolchain: try UserToolchain.default,
+                cacheDir: path,
+                delegate: delegate
+            )
+
+            func check(loader: ManifestLoader, manifest: String) throws {
+                let fileSystem = InMemoryFileSystem()
+                let observability = ObservabilitySystem.makeForTesting()
+
+                let manifestPath = AbsolutePath.root.appending(component: Manifest.filename)
+                try fileSystem.writeFileContents(manifestPath, string: manifest)
+
+                let m = try manifestLoader.load(
+                    manifestPath: manifestPath,
+                    packageKind: .root(.root),
+                    toolsVersion: .current,
+                    fileSystem: fileSystem,
+                    observabilityScope: observability.topScope
+                )
+
+                XCTAssertNoDiagnostics(observability.diagnostics)
+                XCTAssertEqual(m.displayName, "Trivial")
+            }
+
+            do {
+                delegate.prepare()
+                try check(loader: manifestLoader, manifest: manifest)
+                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 1)
+                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 1)
+            }
+
+            do {
+                delegate.prepare(expectParsing: false)
+                try check(loader: manifestLoader, manifest: manifest)
+                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 2)
+                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 1)
+            }
+
+            do {
+                delegate.prepare()
+                try check(loader: manifestLoader, manifest: manifest + "\n\n")
+                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 3)
+                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 2)
+            }
+        }
+    }
+
+    func testCacheInvalidationOnEnv() throws {
+        try testWithTemporaryDirectory { path in
+            let fileSystem = InMemoryFileSystem()
+            let observability = ObservabilitySystem.makeForTesting()
+
+            let manifestPath = path.appending(components: "pkg", "Package.swift")
+            try fileSystem.createDirectory(manifestPath.parentDirectory, recursive: true)
+            try fileSystem.writeFileContents(
+                manifestPath,
+                string: """
+                    import PackageDescription
+                    let package = Package(
+                        name: "Trivial",
+                        targets: [
+                            .target(
+                                name: "foo",
+                                dependencies: []),
+                        ]
+                    )
+                    """
+            )
+
+            let delegate = ManifestTestDelegate()
+
+            let manifestLoader = ManifestLoader(
+                toolchain: try UserToolchain.default,
+                cacheDir: path,
+                delegate: delegate
+            )
+
+            func check(loader: ManifestLoader, expectCached: Bool) throws {
+                delegate.clear()
+                delegate.prepare(expectParsing: !expectCached)
+
+                let manifest = try XCTUnwrap(loader.load(
+                    manifestPath: manifestPath,
+                    packageKind: .root(manifestPath.parentDirectory),
+                    toolsVersion: .current,
+                    fileSystem: fileSystem,
+                    observabilityScope: observability.topScope
+                ))
+
+                XCTAssertNoDiagnostics(observability.diagnostics)
+                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1), [manifestPath])
+                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, expectCached ? 0 : 1)
+                XCTAssertEqual(manifest.displayName, "Trivial")
+                XCTAssertEqual(manifest.targets[0].name, "foo")
+            }
+
+            try check(loader: manifestLoader, expectCached: false)
+            try check(loader: manifestLoader, expectCached: true)
+
+            try withCustomEnv(["SWIFTPM_MANIFEST_CACHE_TEST": "1"]) {
+                try check(loader: manifestLoader, expectCached: false)
+                try check(loader: manifestLoader, expectCached: true)
+            }
+
+            try withCustomEnv(["SWIFTPM_MANIFEST_CACHE_TEST": "2"]) {
+                try check(loader: manifestLoader, expectCached: false)
+                try check(loader: manifestLoader, expectCached: true)
+            }
+
+            try check(loader: manifestLoader, expectCached: true)
+        }
+    }
+
     func testSQLiteCacheHappyCase() throws {
         try testWithTemporaryDirectory { tmpPath in
             let path = tmpPath.appending("test.db")
@@ -41,7 +357,7 @@ class ManifestLoaderCacheTests: XCTestCase {
         }
     }
 
-    func testInMemoryCache() throws {
+    func testInMemoryCacheHappyCase() throws {
         let content = """
             import PackageDescription
             let package = Package(
@@ -53,7 +369,11 @@ class ManifestLoaderCacheTests: XCTestCase {
             )
             """
 
-        let manifestLoader = ManifestLoader(toolchain: try! UserToolchain.default, delegate: .none)
+        let manifestLoader = ManifestLoader(
+            toolchain: try UserToolchain.default,
+            cacheDir: .none,
+            delegate: .none
+        )
 
         let packageURL = "https://scm.com/\(UUID().uuidString)/foo"
 
@@ -75,7 +395,7 @@ class ManifestLoaderCacheTests: XCTestCase {
                 )
             }
             XCTAssertNoDiagnostics(validationDiagnostics)
-            
+
             let deps = Dictionary(uniqueKeysWithValues: manifest.dependencies.map{ ($0.identity.description, $0) })
             XCTAssertEqual(deps["foo"], .remoteSourceControl(url: "https://scm.com/foo", requirement: .upToNextMajor(from: "1.0.0")))
             XCTAssertEqual(deps["bar"], .remoteSourceControl(url: "https://scm.com/bar", requirement: .upToNextMajor(from: "2.1.0")))
@@ -159,9 +479,10 @@ class ManifestLoaderCacheTests: XCTestCase {
 private func makeMockManifests(fileSystem: FileSystem, rootPath: AbsolutePath, count: Int = Int.random(in: 50 ..< 100)) throws -> [ManifestLoader.CacheKey: ManifestLoader.EvaluationResult] {
     var manifests = [ManifestLoader.CacheKey: ManifestLoader.EvaluationResult]()
     for index in 0 ..< count {
-        let manifestPath = rootPath.appending(components: "\(index)", "Package.swift")
+        let packagePath = rootPath.appending("\(index)")
+        let manifestPath = packagePath.appending("Package.swift")
 
-        try fileSystem.createDirectory(manifestPath.parentDirectory, recursive: true)
+        try fileSystem.createDirectory(packagePath, recursive: true)
         try fileSystem.writeFileContents(
             manifestPath,
             string: """
@@ -177,8 +498,8 @@ private func makeMockManifests(fileSystem: FileSystem, rootPath: AbsolutePath, c
             """
         )
         let key = try ManifestLoader.CacheKey(
-            packageIdentity: PackageIdentity(path: manifestPath),
-            packageLocation: manifestPath.pathString,
+            packageIdentity: PackageIdentity(path: packagePath),
+            packageLocation: packagePath.pathString,
             manifestPath: manifestPath,
             toolsVersion: ToolsVersion.current,
             env: [:],
