+++ Tests/PackageLoadingTests/PD_4_2_LoadingTests.swift
@@ -577,212 +577,6 @@ class PackageDescription4_2LoadingTests: PackageDescriptionLoadingTests {
         }
     }
 
-    func testCacheInvalidationOnEnv() throws {
-        try testWithTemporaryDirectory { path in
-            let fs = localFileSystem
-            let observability = ObservabilitySystem.makeForTesting()
-
-            let manifestPath = path.appending(components: "pkg", "Package.swift")
-            try fs.createDirectory(manifestPath.parentDirectory, recursive: true)
-            try fs.writeFileContents(
-                manifestPath,
-                string: """
-                    import PackageDescription
-                    let package = Package(
-                        name: "Trivial",
-                        targets: [
-                            .target(
-                                name: "foo",
-                                dependencies: []),
-                        ]
-                    )
-                    """
-            )
-
-            let delegate = ManifestTestDelegate()
-
-            let manifestLoader = ManifestLoader(toolchain: try UserToolchain.default, cacheDir: path, delegate: delegate)
-
-            func check(loader: ManifestLoader, expectCached: Bool) throws {
-                delegate.clear()
-                delegate.prepare(expectParsing: !expectCached)
-
-                let manifest = try XCTUnwrap(loader.load(
-                    manifestPath: manifestPath,
-                    packageKind: .fileSystem(manifestPath.parentDirectory),
-                    toolsVersion: .v4_2,
-                    fileSystem: fs,
-                    observabilityScope: observability.topScope
-                ))
-
-                XCTAssertNoDiagnostics(observability.diagnostics)
-                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1), [manifestPath])
-                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, expectCached ? 0 : 1)
-                XCTAssertEqual(manifest.displayName, "Trivial")
-                XCTAssertEqual(manifest.targets[0].name, "foo")
-            }
-
-            try check(loader: manifestLoader, expectCached: false)
-            try check(loader: manifestLoader, expectCached: true)
-
-            try withCustomEnv(["SWIFTPM_MANIFEST_CACHE_TEST": "1"]) {
-                try check(loader: manifestLoader, expectCached: false)
-                try check(loader: manifestLoader, expectCached: true)
-            }
-
-            try withCustomEnv(["SWIFTPM_MANIFEST_CACHE_TEST": "2"]) {
-                try check(loader: manifestLoader, expectCached: false)
-                try check(loader: manifestLoader, expectCached: true)
-            }
-
-            try check(loader: manifestLoader, expectCached: true)
-        }
-    }
-
-    func testCaching() throws {
-        try testWithTemporaryDirectory { path in
-            let fs = localFileSystem
-            let observability = ObservabilitySystem.makeForTesting()
-
-            let manifestPath = path.appending(components: "pkg", "Package.swift")
-            try fs.createDirectory(manifestPath.parentDirectory, recursive: true)
-            try fs.writeFileContents(
-                manifestPath,
-                string: """
-                    import PackageDescription
-                    let package = Package(
-                        name: "Trivial",
-                        targets: [
-                            .target(
-                                name: "foo",
-                                dependencies: []),
-                        ]
-                    )
-                    """
-            )
-
-            let delegate = ManifestTestDelegate()
-
-            let manifestLoader = ManifestLoader(toolchain: try UserToolchain.default, cacheDir: path, delegate: delegate)
-
-            func check(loader: ManifestLoader, expectCached: Bool) throws {
-                delegate.clear()
-                delegate.prepare(expectParsing: !expectCached)
-
-                let manifest = try XCTUnwrap(loader.load(
-                    manifestPath: manifestPath,
-                    packageKind: .fileSystem(manifestPath.parentDirectory),
-                    toolsVersion: .v4_2,
-                    fileSystem: fs,
-                    observabilityScope: observability.topScope
-                ))
-
-                XCTAssertNoDiagnostics(observability.diagnostics)
-                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1), [manifestPath])
-                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, expectCached ? 0 : 1)
-                XCTAssertEqual(manifest.displayName, "Trivial")
-                XCTAssertEqual(manifest.targets[0].name, "foo")
-            }
-
-            try check(loader: manifestLoader, expectCached: false)
-            for _ in 0..<2 {
-                try check(loader: manifestLoader, expectCached: true)
-            }
-
-            try fs.writeFileContents(
-                manifestPath,
-                string: """
-                    import PackageDescription
-
-                    let package = Package(
-
-                        name: "Trivial",
-                        targets: [
-                            .target(
-                                name: "foo",
-                                dependencies: [  ]),
-                        ]
-                    )
-
-                    """
-            )
-
-            try check(loader: manifestLoader, expectCached: false)
-            for _ in 0..<2 {
-                try check(loader: manifestLoader, expectCached: true)
-            }
-
-            let noCacheLoader = ManifestLoader(toolchain: try UserToolchain.default, delegate: delegate)
-            for _ in 0..<2 {
-                try check(loader: noCacheLoader, expectCached: false)
-            }
-
-            // Resetting the cache should allow us to remove the cache
-            // directory without triggering assertions in sqlite.
-            manifestLoader.purgeCache(observabilityScope: observability.topScope)
-            XCTAssertNoDiagnostics(observability.diagnostics)
-            try localFileSystem.removeFileTree(path)
-        }
-    }
-
-    func testContentBasedCaching() throws {
-        try testWithTemporaryDirectory { path in
-            let manifest = """
-                import PackageDescription
-                let package = Package(
-                    name: "Trivial",
-                    targets: [
-                        .target(name: "foo"),
-                    ]
-                )
-            """
-
-            let delegate = ManifestTestDelegate()
-
-            let manifestLoader = ManifestLoader(toolchain: try UserToolchain.default, cacheDir: path, delegate: delegate)
-
-            func check(loader: ManifestLoader, manifest: String) throws {
-                let fs = InMemoryFileSystem()
-                let observability = ObservabilitySystem.makeForTesting()
-
-                let manifestPath = AbsolutePath.root.appending(component: Manifest.filename)
-                try fs.writeFileContents(manifestPath, string: manifest)
-
-                let m = try manifestLoader.load(
-                    manifestPath: manifestPath,
-                    packageKind: .root(.root),
-                    toolsVersion: .v4_2,
-                    fileSystem: fs,
-                    observabilityScope: observability.topScope
-                )
-
-                XCTAssertNoDiagnostics(observability.diagnostics)
-                XCTAssertEqual(m.displayName, "Trivial")
-            }
-
-            do {
-                delegate.prepare()
-                try check(loader: manifestLoader, manifest: manifest)
-                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 1)
-                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 1)
-            }
-
-            do {
-                delegate.prepare(expectParsing: false)
-                try check(loader: manifestLoader, manifest: manifest)
-                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 2)
-                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 1)
-            }
-
-            do {
-                delegate.prepare()
-                try check(loader: manifestLoader, manifest: manifest + "\n\n")
-                XCTAssertEqual(try delegate.loaded(timeout: .now() + 1).count, 3)
-                XCTAssertEqual(try delegate.parsed(timeout: .now() + 1).count, 2)
-            }
-        }
-    }
-
     func testProductTargetNotFound() throws {
         let content = """
             import PackageDescription
@@ -980,74 +774,6 @@ class PackageDescription4_2LoadingTests: PackageDescriptionLoadingTests {
             XCTAssertFalse(observability.hasErrorDiagnostics, observability.diagnostics.description)
         }
 #endif
-    }
-
-    final class ManifestTestDelegate: ManifestLoaderDelegate {
-        private let loaded = ThreadSafeArrayStore<AbsolutePath>()
-        private let parsed = ThreadSafeArrayStore<AbsolutePath>()
-        private let loadingGroup = DispatchGroup()
-        private let parsingGroup = DispatchGroup()
-
-        func prepare(expectParsing: Bool = true) {
-            self.loadingGroup.enter()
-            if expectParsing {
-                self.parsingGroup.enter()
-            }
-        }
-
-        func willLoad(packageIdentity: PackageModel.PackageIdentity, packageLocation: String, manifestPath: AbsolutePath) {
-            // noop
-        }
-
-        func didLoad(packageIdentity: PackageIdentity, packageLocation: String, manifestPath: AbsolutePath, duration: DispatchTimeInterval) {
-            self.loaded.append(manifestPath)
-            self.loadingGroup.leave()
-        }
-
-        func willParse(packageIdentity: PackageIdentity, packageLocation: String) {
-            // noop
-        }
-
-        func didParse(packageIdentity: PackageIdentity, packageLocation: String, duration: DispatchTimeInterval) {
-            // noop
-        }
-
-        func willCompile(packageIdentity: PackageIdentity, packageLocation: String, manifestPath: AbsolutePath) {
-            // noop
-        }
-
-        func didCompile(packageIdentity: PackageIdentity, packageLocation: String, manifestPath: AbsolutePath, duration: DispatchTimeInterval) {
-            // noop
-        }
-
-        func willEvaluate(packageIdentity: PackageIdentity, packageLocation: String, manifestPath: AbsolutePath) {
-            // noop
-        }
-
-        func didEvaluate(packageIdentity: PackageIdentity, packageLocation: String, manifestPath: AbsolutePath, duration: DispatchTimeInterval) {
-            self.parsed.append(manifestPath)
-            self.parsingGroup.leave()
-        }
-
-
-        func clear() {
-            self.loaded.clear()
-            self.parsed.clear()
-        }
-
-        func loaded(timeout: DispatchTime) throws -> [AbsolutePath] {
-            guard case .success = self.loadingGroup.wait(timeout: timeout) else {
-                throw StringError("timeout waiting for loading")
-            }
-            return self.loaded.get()
-        }
-
-        func parsed(timeout: DispatchTime) throws -> [AbsolutePath] {
-            guard case .success = self.parsingGroup.wait(timeout: timeout) else {
-                throw StringError("timeout waiting for parsing")
-            }
-            return self.parsed.get()
-        }
     }
 }
